Almost Mana Regeneration
The mana regeneration is almost done, the last this missing from this is a way to stop the regeneration once it has started so you don't have infinite mana while it is regenerating.

+++ Wishing-Star/Assets/Scripts/ItemLibary.cs
@@ -24,7 +24,7 @@ public class ItemLibary : MonoBehaviour
         
     }
 
-    public void ItemLibFind(string tag, Vector2 direction, Vector2 pos, int mana, Collider2D col)
+    public void ItemLibFind(string tag, Vector2 direction, Vector2 pos, float mana, Collider2D col)
     {
         
         switch (tag)
@@ -61,7 +61,7 @@ public class ItemLibary : MonoBehaviour
         
     }
     
-    private void Bow(Vector2 direction, Vector2 pos, int mana, Collider2D col)
+    private void Bow(Vector2 direction, Vector2 pos, float mana, Collider2D col)
     {
